rename VitalsViews to VitalsView
