+++ packages/lesmis/src/views/VitalsView.js
